fixed signed/unsigned printf

+++ Firmware/radio/at.c
@@ -287,7 +287,10 @@ at_i(void)
 		enum ParamID id;
 		// convenient way of showing all parameters
 		for (id = 0; id < PARAM_MAX; id++) {
-			printf("S%d: %s=%d\n", id, param_name(id), (unsigned)param_get(id));
+			printf("S%u: %s=%u\n", 
+			       (unsigned)id, 
+			       param_name(id), 
+			       (unsigned)param_get(id));
 		}
 		tdm_report_timing();
 		return;
